optimise and refactor
