+++ Project.toml
@@ -23,6 +23,7 @@ MetaGraphs = "626554b9-1ddb-594c-aa3c-2596fe9399a5"
 MetaGraphsNext = "fa8bd995-216d-47f1-8a91-f3b68fbeb377"
 OrderedCollections = "bac558e1-5e72-5ebc-8fee-abe8a469f55d"
 Query = "1a8c2f83-1ff3-5112-b086-8aa67b057ba1"
+Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 SintPowerCase = "dcf425a8-86bf-4135-b17a-04ddca2f3b37"
 SintPowerGraphs = "4cf3ad31-9d4e-4532-9fdd-b80a3a06e24f"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
