+++ examples/segmentation/nfc.jl
@@ -13,7 +13,7 @@ function plot_nfc()
     network["nfc", "load"] = NewBranch()
 
     supplies = [vertex for vertex in labels(network) if is_supply(network[vertex])]
-    parts = Set([NetworkPart(network, supply) for supply in supplies])
+    parts = [NetworkPart(network, supply) for supply in supplies]
     optimal_split = segment_network(network, parts)
     display(optimal_split)
     display(plot_that_graph(network, optimal_split))
