+++ src/RelDist.jl
@@ -30,7 +30,7 @@ export MONTH_FACTORS, DAY_FACTORS, HOUR_FACTORS, COST_FUN, LOAD_PROFILES, TEMPER
 include("reconfiguration.jl")
 export calc_R, traverse
 
-include("complete/network_segmenter.jl")
+include("complete/mod.jl")
 export empty_network, Network, Bus, BusKind, NetworkPart, segment_network
 
 include("relrad_calc.jl")
