+++ test/segmentation/simple_overlap.jl
@@ -6,13 +6,13 @@ using Test
 function test_simple_overlap()
     network = empty_network()
 
-    network["bf_6"] = Bus(t_supply, 2.0)
-    network["bf_5"] = Bus(t_supply, 2.0)
+    network["bf_6"] = Bus("bf_6", t_supply, 2.0)
+    network["bf_5"] = Bus("bf_5", t_supply, 2.0)
 
-    network["load_4"] = Bus(t_load, 1.0)
-    network["load_3"] = Bus(t_load, 1.0)
-    network["load_2"] = Bus(t_load, 1.0)
-    network["load_1"] = Bus(t_load, 1.0)
+    network["load_4"] = Bus("load_4", t_load, 1.0)
+    network["load_3"] = Bus("load_3", t_load, 1.0)
+    network["load_2"] = Bus("load_2", t_load, 1.0)
+    network["load_1"] = Bus("load_1", t_load, 1.0)
 
     network["load_1", "load_2"] = NewBranch()
     network["load_2", "load_3"] = NewBranch()
